test(kyc): fix orphaned KycNotStarted test after capabilities rehaul
KycNotStarted.test.tsx imported KycNotStarted + KycRequiresDocuments, both
deleted by the rehaul → module-not-found, suite failed to load. Rewrote as
the callback-wrapper hygiene guard (main's hotfix regression) for the two
surviving states (KycActionRequired, KycFailed) and renamed the file to
match. 2 tests green.
